+++ app/_definitions/meta/cron-jobs.ts
@@ -4,7 +4,7 @@ import kis_sync_audit_status from '../models/cron-jobs/kis_sync_audit_status';
 import kis_sync_base_data_job from '../models/cron-jobs/kis_sync_base_data_job';
 import kis_sync_inventory_data_job from '../models/cron-jobs/kis_sync_inventory_data_job';
 import kis_sync_inventory_notify_job from '../models/cron-jobs/kis_sync_inventory_notify_job';
-import kis_update_tokens_job from '../models/cron-jobs/kis_update_tokens_job';
+import kis_update_tokens_job from '../models/cron-jobs/kisRefreshAccessTokenJob';
 import update_inventory_balance from '../models/cron-jobs/update_inventory_balance';
 
 export default [
