+++ sample.env
@@ -12,7 +12,8 @@ RAPID_LOG_LEVEL=info
 EXPRESS_LOG_DISABLED=false
 
 DISABLE_CRON_JOB=true
-KIS_REFRESH_TOKENS_JOB_DISABLED=true
+KIS_REFRESH_ACCESS_TOKEN_JOB_DISABLED=true
+KIS_REFRESH_AUTH_DATA_JOB_DISABLED=true
 KIS_SYNC_BASE_DATA_JOB_DISABLED=true
 KIS_SYNC_INVENTORY_APPLICATIONS_JOB_DISABLED=true
 WMS_UPDATE_INVENTORY_BALANCE_JOB_DISABLED=true
